run.sh hdfs://jdx-usa  benchmarkRead hdfs://xxx-cluster/user/hive/tmp.orc 20

```
calvinkirs@CalvinKirss-MacBook-Pro hdfs-client-java % bash run.sh hdfs://jdx-usa  benchmarkRead hdfs://xxx-cluster/user/hive/tmp.orc 20

Thread 17 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 261 ms
Thread 9 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00006-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43436 bytes in 289 ms
Thread 11 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 240 ms
Thread 1 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00006-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43436 bytes in 338 ms
Thread 19 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 249 ms
Thread 5 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00006-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43436 bytes in 338 ms
Thread 0 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 390 ms
Thread 13 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00006-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43436 bytes in 205 ms
Thread 14 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 487 ms
Thread 9 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 179 ms
Thread 2 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00006-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43436 bytes in 184 ms
Thread 10 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00006-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43436 bytes in 219 ms
Thread 16 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 502 ms
Thread 8 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 443 ms
Thread 3 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00006-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43436 bytes in 625 ms
Thread 5 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 118 ms
Thread 4 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00005-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 42133 bytes in 387 ms
Thread 2 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 114 ms
Thread 1 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 278 ms
Thread 13 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 183 ms
Thread 7 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00005-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 42133 bytes in 1236 ms
Thread 3 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 130 ms
Thread 10 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 186 ms
Thread 4 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00006-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43436 bytes in 135 ms
Thread 7 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00006-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43436 bytes in 150 ms
Thread 4 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 132 ms
Thread 7 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 133 ms
Thread 15 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 1374 ms
Thread 12 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00006-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43436 bytes in 1353 ms
Thread 12 read file hdfs://jdx-usa/user/hive/tmp.orc/part-00007-18044f6c-c750-41db-8a21-e1f949c14105-c000.zlib.orc: 43203 bytes in 183 ms
All threads completed. Results:
Thread 8 done: 9 files, 343077 bytes, 0 failures, total time 2032 ms

Thread 18 done: 9 files, 343077 bytes, 0 failures, total time 1771 ms

Thread 15 done: 9 files, 343077 bytes, 0 failures, total time 3035 ms

Thread 13 done: 9 files, 343077 bytes, 0 failures, total time 2159 ms

Thread 1 done: 9 files, 343077 bytes, 0 failures, total time 2157 ms

Thread 12 done: 9 files, 343077 bytes, 0 failures, total time 3220 ms

Thread 3 done: 9 files, 343077 bytes, 0 failures, total time 2181 ms

Thread 11 done: 9 files, 343077 bytes, 0 failures, total time 1873 ms

Thread 5 done: 9 files, 343077 bytes, 0 failures, total time 2093 ms

Thread 6 done: 9 files, 343077 bytes, 0 failures, total time 1675 ms

Thread 9 done: 9 files, 343077 bytes, 0 failures, total time 2007 ms

Thread 14 done: 9 files, 343077 bytes, 0 failures, total time 1995 ms

Thread 2 done: 9 files, 343077 bytes, 0 failures, total time 2122 ms

Thread 17 done: 9 files, 343077 bytes, 0 failures, total time 1825 ms

Thread 19 done: 9 files, 343077 bytes, 0 failures, total time 1930 ms

Thread 0 done: 9 files, 343077 bytes, 0 failures, total time 1980 ms

Thread 7 done: 9 files, 343077 bytes, 0 failures, total time 2451 ms

Thread 4 done: 9 files, 343077 bytes, 0 failures, total time 2362 ms

Thread 16 done: 9 files, 343077 bytes, 0 failures, total time 2010 ms

Thread 10 done: 9 files, 343077 bytes, 0 failures, total time 2197 ms

TOTAL: 180 files, 6861540 bytes read, 0 failed
HDFS client closed


```